Fix: Perbaikan konfigurasi PHPUnit dan error export Excel corrupt

+++ vehicle-booking/app/Database/Migrations/2026-08-19-100903_AddImageUrlToVehicles.php
@@ -20,6 +20,9 @@ class AddImageUrlToVehicles extends Migration
 
     public function down()
     {
-        $this->forge->dropColumn('vehicles', 'image_url');
+        // Pengecekan agar SQLite3 tidak error saat mencoba dropColumn pada testing
+        if ($this->db->DBDriver !== 'SQLite3') {
+            $this->forge->dropColumn('vehicles', 'image_url');
+        }
     }
 }
